pyproject.toml
[tool.poetry]
name = "foldwrap_backend"
version = "0.1.0"
description = ""
authors = ["Ress <ress@foldwrap.com>"]

[tool.poetry.scripts]
start = "app:main"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
dnspython = "^2.4.2"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
filetype = "^1.2.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
pymongo = "^4.5.0"
motor = "^3.3.2"
pydantic = {extras = ["email"], version = "^2.4.0"}
pillow = "^10.1.0"
requests = "^2.31.0"
fastapi-restful = "^0.5.0"
pydantic-settings = "^2.1.0"
typing-inspect = "^0.9.0"
structlog = "^23.2.0"
authlib = "^1.2.1"
loguru = "^0.7.2"
httpx = "^0.25.1"
websockets = "^12.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.32.0"}
redis = "^5.0.1"
fonttools = "^4.44.3"




[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
coverage = "^7.3.1"
flake8 = "^6.1.0"
oitnb = "^0.2.2"
black = "^23.9.1"
pytest-asyncio = "^0.21.1"
ruff = "^0.0.291"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# formatting config
[tool.black]
line-length = 220
extend-exclude = '''
/(
  | .env
  | .venv
  | venv
  | logs
  | uploads
  | testing_assets
)/
'''

[tool.ruff]
# show-source = true
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".ruff_cache"
]
# Same as Black.
line-length = 280
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Assume Python 3.12.
target-version = "py312"
task-tags = ["E501"]

